@@ -5,14 +5,14 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   file_system_explorer:
-    path: ..\
+    path: ../
 
 dev_dependencies:
   flutter_test:
