@@ -21,16 +21,16 @@ class MyApp extends StatelessWidget {
             color: Color(0xffbbbbbb),
           ),
           textTheme: TextTheme(
-            body1: TextStyle(color: Color(0xffbbbbbb)),
+            bodyMedium: TextStyle(color: Color(0xffbbbbbb)),
           )
       ),
-      home: new MyHomePage(),
+      home: new MyHomePage(key: ValueKey("homepage")),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key}) : super(key: key);
+  MyHomePage({required Key key}) : super(key: key);
 
   @override
   _MyHomePageState createState() => new _MyHomePageState();
@@ -38,7 +38,7 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
 
-  String path;
+  String? path;
 
   @override
   Widget build(BuildContext context) {
